services/obsidian-bridge/requirements.txt
fastapi
uvicorn[standard]
paho-mqtt>=2.0
loguru
python-dotenv
watchdog
